@@ -8,14 +8,14 @@ const BorderList = ({ borders, theme }) => {
   const [borderCountries, setBorderCountries] = useState([]);
   useEffect(() => {
     axios
-      .get(`https://restcountries.eu/rest/v2/alpha?codes=${borders.join(";")}`)
-      .then(res => setBorderCountries(res.data))
-      .catch(err => console.log(err));
+      .get(`https://restcountries.com/v2/alpha?codes=${borders.join(",")}`)
+      .then((res) => setBorderCountries(res.data))
+      .catch((err) => console.log(err));
   }, [borders]);
 
   return (
     <>
-      {borderCountries.map(country => (
+      {borderCountries.map((country) => (
         <Link
           className={theme === "Light" ? styles.border : styles.border_dark}
           to={`/country/${country.alpha3Code}`}
