@@ -1,7 +1,8 @@
-import React, { useEffect, useContext } from "react";
+import React, { useEffect, useContext, useState } from "react";
 import { Link } from "react-router-dom";
+import axios from "axios";
 
-import { GlobalContext } from "../context/GlobalState";
+import { GlobalContext } from "../../context/GlobalState";
 import BorderList from "./BorderList";
 import Info from "./Info";
 import ExtraInfo from "./ExtraInfo";
@@ -9,10 +10,24 @@ import ExtraInfo from "./ExtraInfo";
 import styles from "./CountryDetails.module.css";
 
 const CountryDetails = ({ match }) => {
-  const { country, fetchCountry, theme } = useContext(GlobalContext);
+  const { setLoading, theme } = useContext(GlobalContext);
+  const [country, setCountry] = useState({});
+
   const countryCode = match.params.code;
 
   useEffect(() => {
+    //  Get single country details
+    const fetchCountry = (code) => {
+      setLoading(true);
+      axios
+        .get(`https://restcountries.com/v2/alpha/${code}`)
+        .then((res) => {
+          setCountry(res.data);
+          setLoading(false);
+        })
+        .catch((err) => console.log(err));
+    };
+
     fetchCountry(countryCode);
   }, [countryCode]);
 
@@ -27,7 +42,7 @@ const CountryDetails = ({ match }) => {
     borders = [],
     topLevelDomain = [],
     currencies = [],
-    languages = []
+    languages = [],
   } = country;
 
   return (
@@ -36,7 +51,22 @@ const CountryDetails = ({ match }) => {
         to="/"
         className={theme === "Light" ? styles.back : styles.back_dark}
       >
-        <ion-icon name="arrow-round-back" />
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          strokeWidth="2"
+          stroke="currentColor"
+          fill="none"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        >
+          <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+          <line x1="5" y1="12" x2="19" y2="12"></line>
+          <line x1="5" y1="12" x2="9" y2="16"></line>
+          <line x1="5" y1="12" x2="9" y2="8"></line>
+        </svg>
         Back
       </Link>
       <div className={styles.details}>
