@@ -3,10 +3,10 @@ import React from "react";
 import styles from "./CountryList.module.css";
 
 const CountryCard = ({ details }) => {
-  const { flag, name, population, region, capital } = details;
+  const { flags, name, population, region, capital } = details;
   return (
     <div className={styles.country_card}>
-      <img src={flag} alt={`${name}'s Flag`} />
+      <img src={flags.svg} alt={`${name}'s Flag`} />
       <div className={styles.info}>
         <p className={styles.country_name}>{name}</p>
         <p>
