@@ -1,17 +1,13 @@
 import React, { useContext } from "react";
-import { GlobalContext } from "../context/GlobalState";
+import { GlobalContext } from "../../context/GlobalState";
 
 import styles from "./Filter.module.css";
 
-const Search = () => {
-  const { filterByRegion, fetchCountries, theme } = useContext(GlobalContext);
-  const onUserInput = e => {
+const Search = ({ setRegion }) => {
+  const { theme } = useContext(GlobalContext);
+  const onUserInput = (e) => {
     let value = e.target.value;
-    if (!value) {
-      fetchCountries();
-    } else {
-      filterByRegion(value);
-    }
+    setRegion(value);
   };
 
   return (
@@ -19,8 +15,11 @@ const Search = () => {
       <select
         onChange={onUserInput}
         className={theme === "Light" ? "" : styles.dark}
+        defaultValue="title"
       >
-        <option disabled>Filter by Region</option>
+        <option disabled value="title">
+          Filter by Region
+        </option>
         <option value="">All</option>
         <option value="africa">Africa</option>
         <option value="americas">Americas</option>
@@ -28,7 +27,18 @@ const Search = () => {
         <option value="europe">Europe</option>
         <option value="oceania">Oceania</option>
       </select>
-      <ion-icon name="arrow-dropdown" />
+
+      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
+        <title>Chevron Down</title>
+        <path
+          fill="none"
+          stroke="currentColor"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          strokeWidth="48"
+          d="M112 184l144 144 144-144"
+        />
+      </svg>
     </div>
   );
 };
